# openapi.yaml
openapi: 3.0.1
info:
  title: AIHealth API
  description: This is a AIHealth server.
  contact: {}
  version: "1.0.0"
servers:
- url: http://localhost:10086
...
paths:
...
  /books:
    get:
      tags:
      - "books"
      description: View books detail information
      responses:
        200:
          description: ""
          headers:
            books:
              description: Books
              schema:
                type: string
          content: {}
    post:
      tags:
      - "books"
      description: Add books detail information
      requestBody:
        description: Add Book
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/model.Book'
        required: true
      responses:
        200:
          description: ""
          content: {}
      x-codegen-request-body-name: book
  /books/{isbn}:
    delete:
      tags:
      - "books"
      description: Delete book by isbn
      parameters:
      - name: isbn
        in: path
        description: ISBN
        required: true
        schema:
          type: string
      responses:
        200:
          description: ""
          content: {}
    patch:
      tags:
      - "books"
      description: Update book by isbn
      parameters:
      - name: isbn
        in: path
        description: Book Id
        required: true
        schema:
          type: string
      requestBody:
        description: Add Book
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/model.Book'
        required: true
      responses:
        200:
          description: ""
          content: {}
      x-codegen-request-body-name: book
  /books/isbn/{isbn}:
    get:
      tags:
      - "books"
      description: View books detail information with Book ID
      parameters:
      - name: isbn
        in: path
        description: Book ISBN
        required: true
        schema:
          type: string
      responses:
        200:
          description: ""
          headers:
            books:
              description: Books
              schema:
                type: string
          content: {}
...
components:
  schemas:
    model.Book:
      type: object
      properties:
        author:
          type: string
        isbn:
          type: string
        language:
          type: string
        publisher:
          type: string
        rating:
          type: integer
        tags:
          type: string
        title:
          type: string
          example: Go in Action
        year:
          type: string
...